+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNP589GWH2RWC9H81P2RA538"
+        "01KNPB0DJ7BM380QCGEGXNH023"
     )
     r"""The unique identifier of the data source"""