+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNP589GZ5AW65KEM9V40EZ2B"
+        "01KNPB0DJA3FWMXNWX8G192P7N"
     )
     r"""The unique identifier of the data source"""