+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-08T09:02:10.655Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-08T10:42:44.319Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")