+++ packages/orq-rc/RELEASES.md
@@ -6929,3 +6929,13 @@ Based on:
 - [python v4.7.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.53 - packages/orq-rc
+
+## 2026-04-08 10:45:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.54] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.54 - packages/orq-rc